Rename file temporarily to luppy; implement much of inotify handling
The current design is to recur through the current folder, watching
all lup.cl files and any directory. Adding a lup.cl file to a directory
should immediately be loaded and executed. Changing a lup.cl file should
reload it.

There must be a table of dependencies. For example we can create a simple hash
table like so:

	(8 (1 2 3 4 ...))

Where the numbers (1 2 3 4 ...) are the files which depend on this one (8).
This can be used because we always get an identifier from inotify. On the other
hand we also need to store filenames in another table:

	(8 ("some_file.c" ("command" I O) "some_file.o))

Simply to keep track of how and what to compile.

So we get a number from inotify. We look it up in the table, execute the command,
then look it up in the dependents, execute for each of those recursively.

So we split the problem up, first I need a function for executing the commands, that
would be very useful. It also needs to find out how to store the result in the
output tree.

What about includes? Oh almost forgot

	(8 ("subdir/here/there/" ("include1.h" "include2.h" ..) ("some_file.c") ("command" I O) ("some_file.o")))

So we've got subdirectory, auxiliary dependencies, main dependencies, and output files.
If necessary the user can write (aref I 0) or something. The lup.cl will look something like this:

	(
		(batch "*.h" ("my_main.c" "my_lib.c") ("gcc" I "-o" O) "out.o")
		(single "boost/or/something/*.h" ("first_main.c" "second_lib.c") ("gcc" "-c" I "-o" O) ("first_lib.o" "second_lib.o"))
	)

We see that the DSL is quite robust, any element can be turned into a list so youcan specify more files.
Another thing is that the commands should be checked for accessing elements during runtime. It'd make
the whole process so much easier; you know, not having to specify stuff.
It can be done using inotify by attacking watches to EVERYTHING, but that seems a bit much.
Maybe it's better to specify it explicitly? Who knows...

What about globbing?

Another idea is to just write the commands and let lup figure out what's dependent and what's created.
How do we allow toolchain and customization then? Say if we want to build for a specific toolchain like
qemu but not gcc? Hmmm,... we could set variables when the system is running...

You know, like you set a variable whilst the program is running, and you use it inside the lup.cl files:

	(toolchain "*.cpp")

Since there are watches, anything that's generated by the command will be moved to output/.
Setting (setf #'toolchain #'gnu-linux-gcc-whatever) could change the toolchain. But there
are many more targets to handle... this is a tought problem. I'll have to think about it.